+++ emoji/__init__.py
@@ -21,7 +21,7 @@ __all__ = [
     'LANGUAGES',
 ]
 
-__version__ = '2.14.0'
+__version__ = '2.14.1'
 __author__ = 'Taehoon Kim, Kevin Wurster'
 __email__ = 'carpedm20@gmail.com'
 # and wursterk@gmail.com, tahir.jalilov@gmail.com
@@ -29,7 +29,7 @@ __source__ = 'https://github.com/carpedm20/emoji/'
 __license__ = """
 New BSD License
 
-Copyright (c) 2014-2024, Taehoon Kim, Kevin Wurster
+Copyright (c) 2014-2025, Taehoon Kim, Kevin Wurster
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without